feature 13221 modified: merge to trunk
git-svn-id: http://talendforge.org/svn/tos/trunk@45911 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.core/src/main/java/org/talend/core/ui/metadata/editor/MetadataEmfTableEditorView.java
@@ -169,7 +169,7 @@ public class MetadataEmfTableEditorView extends AbstractMetadataTableEditorView<
                 // }
                 // }
                 // }
-                return bean.getPrecision();
+                return Long.valueOf(bean.getPrecision()).intValue();
             }
 
             public void set(MetadataColumn bean, Integer value) {
@@ -198,7 +198,7 @@ public class MetadataEmfTableEditorView extends AbstractMetadataTableEditorView<
                 // }
                 // }
                 // }
-                return bean.getLength();
+                return Long.valueOf(bean.getLength()).intValue();
             }
 
             public void set(MetadataColumn bean, Integer value) {
